hide immutables from the API
the immutables library ends up being exposed through the API if you do not define the Builder interface for a specific class. this means you are requiring consuming applications to include the library or they'll throw an exception at build time.

the solution is to hide the implementation class using the pattern described in the immutables docs here: http://immutables.github.io/immutable.html#hide-implementation-class